src/enhancement/clahe.py
import cv2
import numpy as np

def clahe_enhance(image: np.ndarray, clip_limit: float = 2.0, grid_size: tuple = (8, 8)) -> np.ndarray:
    """
    Apply Contrast Limited Adaptive Histogram Equalization (CLAHE) to enhance visibility.
    Useful for fog, low light, and thermal imagery.
    """
    if image is None:
        return None

    # Convert to LAB color space
    lab = cv2.cvtColor(image, cv2.COLOR_BGR2LAB)
    l, a, b = cv2.split(lab)

    # Apply CLAHE to L-channel
    clahe = cv2.createCLAHE(clipLimit=clip_limit, tileGridSize=grid_size)
    cl = clahe.apply(l)

    # Merge channels and convert back to BGR
    limg = cv2.merge((cl, a, b))
    enhanced = cv2.cvtColor(limg, cv2.COLOR_LAB2BGR)

    return enhanced

def dark_channel_prior(image: np.ndarray) -> np.ndarray:
    """
    Placeholder for Dark Channel Prior based dehazing.
    """
    # Simplified placeholder implementation
    return clahe_enhance(image, clip_limit=3.0)
