requirements.txt
# Core dependencies
streamlit>=1.28.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0

# Deep Learning
torch>=2.0.0; sys_platform != "win32"
torchvision>=0.15.0; sys_platform != "win32"
ultralytics>=8.0.0

# ONNX and Runtime
onnx>=1.14.0
onnxruntime>=1.15.0
# onnxruntime-gpu>=1.15.0  # Uncomment for GPU environments

# Computer Vision
scikit-image>=0.21.0
scipy>=1.11.0

# Tracking
filterpy>=1.4.5
scikit-learn>=1.3.0

# System Monitoring
psutil>=5.9.0
# GPUtil>=1.4.0  # Often fails in containerized CPU environments

# Data Processing
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Video Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Optional: Database
# pymongo>=4.5.0

# Optional: API
# fastapi>=0.103.0
# uvicorn>=0.23.0
